pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "printitbot"
version = "0.1.0"
description = "Печатальный бот"
authors = [
    {name = "Boris", email = "chechko_boris@mail.ru"}
]
dependencies = [
    "requests",
    "python-telegram-bot"
]
requires-python = ">=3.9"

[project.scripts]
printitbot = "src.main:main"
